+++ R/set_attr.R
@@ -2,8 +2,11 @@
 #'
 #' \itemize{
 #'   \item character vectors collapsed to a space-delimited character string
-#'   \item logicals are converted to "true" or "false"
+#'   \item logicals are converted to `"true"` or `"false"`
 #'   \item lubridate durations are converted to numeric (milliseconds)
+#'   \item lists are converted to JSON,
+#'     **Limitation**: [lubridate::duration()] objects within lists
+#'     will not translate correctly
 #' }
 #'
 #' @param x  value to be converted
@@ -43,6 +46,17 @@ bs_attr.logical <- function(x){
 }
 
 #' @rdname bs_attr
+#' @rawNamespace S3method(bs_attr,list)
+#' @keywords internal
+#' @export
+bs_attr.list <- function(x){
+
+  x <- jsonlite::toJSON(x, auto_unbox = TRUE)
+
+  x
+}
+
+#' @rdname bs_attr
 #' @rawNamespace S3method(bs_attr,Duration)
 #' @keywords internal
 #' @export
@@ -65,6 +79,12 @@ setGeneric("bs_attr", useAsDefault = bs_attr.default)
 #' @export
 #'
 setMethod("bs_attr", list("logical"), bs_attr.logical)
+
+#' @rdname bs_attr
+#' @keywords internal
+#' @export
+#'
+setMethod("bs_attr", list("list"), bs_attr.list)
 
 #' @rdname bs_attr
 #' @keywords internal
@@ -106,7 +126,7 @@ bs_set_attr <- function(tag, .prefix = "data", ...){
   names_to_keep <- setdiff(names_existing, names_new)
 
   # keep those attributes
-  tag$attribs <- tag$attribs[names_to_keep]
+  tag$attribs <- tag$attribs[names_existing %in% names_to_keep]
 
   # append these attributes to the tag
   args <- c(list(tag = tag), attributes_bs)
@@ -121,12 +141,12 @@ bs_set_attr <- function(tag, .prefix = "data", ...){
 #'
 #' One of the mechanisms used by the API for Boostrap JavaScript-components is
 #' an html elements' attributes. These attribute names are prefixed with
-#' \code{"data-"} or \code{"aria-"}, depending on the function.
+#' `"data-"` or `"aria-"`, depending on the function.
 #'
 #' When expressed in html, attributes themselves have the properties:
 #'
 #' \itemize{
-#'   \item Logical values are expressed as \code{"true"} or \code{"false"}.
+#'   \item Logical values are expressed as `"true"` or `"false"`.
 #'   \item Time durations are expressed as number of milliseconds.
 #'   \item Vector (non scalar) values are expressed in a space-delimited list.
 #' }
@@ -135,17 +155,17 @@ bs_set_attr <- function(tag, .prefix = "data", ...){
 #' familiar to you as an R user. For example:
 #'
 #' \itemize{
-#'   \item Logical values can be expressed as logicals: \code{TRUE} or \code{FALSE}.
+#'   \item Logical values can be expressed as logicals: `TRUE` or `FALSE`.
 #'   \item Time durations can be expressed using lubridate durations.
 #'   \item Vector (non scalar) values can be expressed as vectors.
 #' }
 #'
 #' Note that this returns a modified copy of the tag sent to it, so it is pipeable.
 #'
-#' @param tag     \code{htmltools::\link[htmltools]{tag}}
-#' @param ...     named arguments used to set the attributes of \code{tag}
+#' @param tag     `htmltools::[tag][htmltools::tag]`
+#' @param ...     named arguments used to set the attributes of `tag`
 #'
-#' @return \code{htmltools::\link[htmltools]{tag}}, modified copy of \code{tag}
+#' @return `htmltools::[tag][htmltools::tag]`, modified copy of `tag`
 #' @examples
 #' library("htmltools")
 #' library("lubridate")
@@ -159,7 +179,7 @@ bs_set_attr <- function(tag, .prefix = "data", ...){
 #'   bs_set_aria(expanded = FALSE)
 #' @export
 #'
-#' @seealso \href{http://getbootstrap.com/javascript}{Bootstrap JavaScript Components}
+#' @seealso [Bootstrap JavaScript Components](http://getbootstrap.com/docs/3.3/javascript)
 #
 bs_set_data <- function(tag, ...){
   bs_set_attr(tag, .prefix = "data", ...)